pyproject.toml
[project]
name = "smartmeetingroom_leilamounzer"
version = "0.1.0"
description = "Smart Meeting Room & Management System backend services"
authors = [{ name = "Leila Mounzer" }]
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.23",
  "sqlalchemy>=2.0",
  "psycopg2-binary>=2.9",
  "alembic>=1.12",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "python-multipart>=0.0.6",
  "pydantic[email]>=2.6",
  "pydantic-settings>=2.2",
  "gunicorn>=21.2",
  "httpx>=0.27",
  "requests>=2.31",
  "slowapi>=0.1.8",
  "cachetools>=5.3",
  "pika>=1.3",
  "circuitbreaker>=1.5"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "tox>=4.11",
  "black>=23.11",
  "ruff>=0.1",
  "sphinx>=7.2",
  "sphinx-rtd-theme>=1.3",
  "memory-profiler>=0.61",
  "locust>=2.19"
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
  "tests"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
